requirements.txt
discord.py
python-dotenv
ics
requests
aiohttp